Update ingress.yaml with new annotations
New annotations for cert-manager and nginx were added in the 'ingress.yaml' file. The 'acme.cert-manager.io/http01-edit-in-place' enables in-place authorization, while 'nginx.ingress.kubernetes.io/auth-tls-pass-certificate-to-upstream' ensures that the certificate is passed to the upstream server. Both these changes allow for more flexibility and better SSL/TLS management.